# spacecraft-infrastructure/src/main/java/com/w2m/spacecraft/infrastructure/rest/dto/SpacecraftRQDTO.java
package com.w2m.spacecraft.infrastructure.rest.dto;

import lombok.Data;
import lombok.EqualsAndHashCode;

@EqualsAndHashCode(callSuper = true)
@Data
public class SpacecraftRQDTO extends SpacecraftRDTO {

}